+++ extensions/hevy/src/search-routines.tsx
@@ -1,7 +1,6 @@
 import { List } from "@raycast/api";
 import { usePromise } from "@raycast/utils";
 import { getRoutines } from "./lib/api";
-import RoutineListItem from "./components/RoutineListItem";
 
 export default function Command() {
   const { data, isLoading } = usePromise(() => getRoutines());
@@ -9,7 +8,12 @@ export default function Command() {
   return (
     <List isLoading={isLoading}>
       {data?.routines.map((routine) => (
-        <RoutineListItem key={routine.id} routine={routine} />
+        <List.Item
+          key={routine.id}
+          title={routine.title}
+          subtitle={routine.created_at}
+          accessories={[{ text: `${routine.exercises.length} exercises` }]}
+        />
       ))}
     </List>
   );